// apps/people/skill-set/browser/src/pages/CladogramPage/entity.schema.types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ICladogramNode {
  /**
   * The name of the node.
   */
  name: string;
  /**
   * The length property of the node.
   */
  length: number;
  /**
   * A number between 1 and 5 representing the level expertise in the topic.
   */
  stars?: number;
  /**
   * A detailed description of what is included in this skill.
   */
  description?: string | null;
  /**
   * Array of child nodes.
   */
  branchset: ICladogramNode[];
}